requirements.txt
# ============================================================
# requirements.txt
# Top-level project dependencies
# For development, testing, and CI
# ============================================================

# ── Deep learning ──────────────────────────────────────────
torch==2.2.2
torchvision==0.17.2

# ── MLOps ─────────────────────────────────────────────────
mlflow==2.12.1
dvc==3.49.0

# ── Data processing ────────────────────────────────────────
pillow==10.3.0
pandas==2.2.2
scikit-learn==1.4.2
numpy==1.26.4

# ── API ────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.0
prometheus-client==0.20.0

# ── Pipeline ───────────────────────────────────────────────
apache-airflow==2.8.1
pyyaml==6.0.1

# ── Testing ────────────────────────────────────────────────
pytest==8.1.1
pytest-cov==5.0.0
httpx==0.27.0

# ── Code quality ───────────────────────────────────────────
flake8==7.0.0
